Extract product analytics monitoring
